Make fonts smaller, update dependencies
I had trouble starting locally because the old command `bundle exec
jekyll liveserve` was actually part of an outdated plugin. That
functionality has since been added to jekkyl (https://github.com/jekyll/jekyll/issues/9451).
As part of my faffing around with that I also upgraded other stuff.

I've added in alpine.js as a dependency to make things easily show/hide
on the research page. It's syntax is a little funky but it's convenient
for stuff like this and it is quite simple.